Added products pagination

+++ server/src/controllers/productsControllers.ts
@@ -13,8 +13,6 @@ export const getAllProducts = async (_req: express.Request, res: express.Respons
 export const createProduct = async (req: any, res: express.Response) => {
   const product = req.body
 
-  console.log(req.file)
-
   try {
     product.imageUrl = 'uploads/' + req.file.filename
     product.userId = req.user.userId
@@ -32,7 +30,6 @@ export const updateProduct = (req: express.Request, res: express.Response) => {
   try {
     productsServices.updateProduct(product)
     res.status(200).send('Product updated')
-   
   } catch (error) {
     res.status(400).send(error)
   }
@@ -40,14 +37,28 @@ export const updateProduct = (req: express.Request, res: express.Response) => {
 export const deleteProduct = async (req: express.Request, res: express.Response) => {
   const { productId } = req.body
   try {
-   const productDeleted = await productsServices.deleteProduct(productId)
+    const productDeleted = await productsServices.deleteProduct(productId)
 
-   fs.rm(productDeleted.imageUrl, (error) => {
-    console.log(error);
-   })
+    if (productDeleted) {
+      fs.rm(productDeleted.imageUrl, (error) => {
+        console.log(error)
+      })
+    }
 
-    res.status(200).send({ message: 'Product deleted', productDeleted})
+    res.status(200).send({ message: 'Product deleted', productDeleted })
   } catch (error) {
     res.status(400).send(error)
+  }
+}
+
+export const getProductsByCategory = async (req: express.Request, res: express.Response) => {
+  const { categoryId, cursor } = req.query
+  try {
+    if (categoryId) {
+      const productsByCategory = await productsServices.getProductsByCategory(categoryId, cursor)
+      res.status(200).send(productsByCategory)
+    }
+  } catch (error: any) {
+    res.send(error.message)
   }
 }
